coveo-mcp-server/requirements.txt
mcp>=1.10.0
boto3>=1.26.0
httpx>=0.25.0
python-dotenv>=0.19.0
bedrock-agentcore
asyncio
dataclasses
typing-extensions
